# src/main/java/ru/skypro/homework/model/User.java
package ru.skypro.homework.model;

import com.fasterxml.jackson.annotation.JsonProperty;
import jakarta.persistence.*;
import lombok.*;
import ru.skypro.homework.dto.Role;

import java.util.List;

@Entity
@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class User {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    @JsonProperty(access = JsonProperty.Access.READ_ONLY)
    private Integer id;

    private String email;
    private String firstName;
    private String lastName;
    private String phone;
    private Role role;

    @OneToOne(mappedBy = "user", cascade = CascadeType.ALL)
    private UserImage image;

    @OneToMany(fetch = FetchType.LAZY, cascade = CascadeType.ALL, mappedBy = "user")
    private List<Ad> ads;
}